lsp: moved to cmp since more completions stuff is in there
